+++ src/codec/encoders/lossless/LosslessIntra.cpp
@@ -21,7 +21,7 @@ void LosslessIntraEncoder::encode() {
         double sum = 0;
         const auto sample_f = sample_frames(frames, sample_factor);
         for (auto &frame: sample_f) { sum += Golomb::adjust_m(frame->get_intra_encoding()); }
-        const int k = static_cast<int>(sum / sample_f.size());
+        const int k = static_cast<int>(sum / static_cast<double>(sample_f.size()));
         const int golomb_m = 1 << k;
         this->golomb_m = golomb_m;
     }
